Adds tag class views
