+++ bookwyrm/tests/test_view_actions.py
@@ -48,8 +48,6 @@ class ViewActions(TestCase):
         self.factory = RequestFactory()
 
 
-
-
     def test_edit_shelf_privacy(self):
         ''' set name or privacy on shelf '''
         shelf = self.local_user.shelf_set.get(identifier='to-read')
@@ -168,43 +166,3 @@ class ViewActions(TestCase):
         self.assertEqual(readthrough.finish_date.day, 7)
         self.assertEqual(readthrough.book, self.book)
         self.assertEqual(readthrough.user, self.local_user)
-
-
-    def test_tag(self):
-        ''' add a tag to a book '''
-        request = self.factory.post(
-            '', {
-                'name': 'A Tag!?',
-                'book': self.book.id,
-            })
-        request.user = self.local_user
-
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            actions.tag(request)
-
-        tag = models.Tag.objects.get()
-        user_tag = models.UserTag.objects.get()
-        self.assertEqual(tag.name, 'A Tag!?')
-        self.assertEqual(tag.identifier, 'A+Tag%21%3F')
-        self.assertEqual(user_tag.user, self.local_user)
-        self.assertEqual(user_tag.book, self.book)
-
-
-    def test_untag(self):
-        ''' remove a tag from a book '''
-        tag = models.Tag.objects.create(name='A Tag!?')
-        models.UserTag.objects.create(
-            user=self.local_user, book=self.book, tag=tag)
-        request = self.factory.post(
-            '', {
-                'user': self.local_user.id,
-                'book': self.book.id,
-                'name': tag.name,
-            })
-        request.user = self.local_user
-
-        with patch('bookwyrm.broadcast.broadcast_task.delay'):
-            actions.untag(request)
-
-        self.assertTrue(models.Tag.objects.filter(name='A Tag!?').exists())
-        self.assertFalse(models.UserTag.objects.exists())
