+++ bookwyrm/view_actions.py
@@ -9,7 +9,6 @@ from django.utils import timezone
 from django.views.decorators.http import require_POST
 
 from bookwyrm import forms, models, outgoing
-from bookwyrm.broadcast import broadcast
 from bookwyrm.vviews import get_user_from_username, get_edition
 
 @login_required
@@ -213,47 +212,6 @@ def create_readthrough(request):
         return redirect(book.local_path)
     readthrough.save()
     return redirect(request.headers.get('Referer', '/'))
-
-
-@login_required
-@require_POST
-def tag(request):
-    ''' tag a book '''
-    # I'm not using a form here because sometimes "name" is sent as a hidden
-    # field which doesn't validate
-    name = request.POST.get('name')
-    book_id = request.POST.get('book')
-    book = get_object_or_404(models.Edition, id=book_id)
-    tag_obj, created = models.Tag.objects.get_or_create(
-        name=name,
-    )
-    user_tag, _ = models.UserTag.objects.get_or_create(
-        user=request.user,
-        book=book,
-        tag=tag_obj,
-    )
-
-    if created:
-        broadcast(request.user, user_tag.to_add_activity(request.user))
-    return redirect('/book/%s' % book_id)
-
-
-@login_required
-@require_POST
-def untag(request):
-    ''' untag a book '''
-    name = request.POST.get('name')
-    tag_obj = get_object_or_404(models.Tag, name=name)
-    book_id = request.POST.get('book')
-    book = get_object_or_404(models.Edition, id=book_id)
-
-    user_tag = get_object_or_404(
-        models.UserTag, tag=tag_obj, book=book, user=request.user)
-    tag_activity = user_tag.to_remove_activity(request.user)
-    user_tag.delete()
-
-    broadcast(request.user, tag_activity)
-    return redirect('/book/%s' % book_id)
 
 
 @login_required
